server/livekit_agent/requirements.txt
livekit
livekit-agents[codecs]
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
